feat: restructure user components by moving UserRow and UserMetadata to the user directory
